基本上可以正常遊玩，想到在加功能 目前未發現明顯bug

+++ src/hang_up_game/java/game/Background.java
@@ -155,7 +155,7 @@ public class Background {
 		}
 	}
 	private static int getDelay(int strong) {
-		return 1000 / (strong * 100);
+		return (int)(1000.0 / strong * 100);
 	}
 	public static Set<MachineMiner> getAllMinerOnline() {
 		return machines.keySet();
